controlar accesos indeseables

+++ app/controllers/procesarBorrarTarea.php
@@ -5,11 +5,25 @@
          * @param  string $id id de la tarea
          */
 
-         require __DIR__ . '/../ctes.php';
-        include(MODEL_PATH.'GestionDataBase.php'); 
-        include(MODEL_PATH.'Tarea.php'); 
+        session_start();
+        if($_SESSION['rol'] == 'Administrador'){
 
-        $id = $_GET['id'];
-        Tarea::deleteTarea($id);
+                require __DIR__ . '/../ctes.php';
+                include(MODEL_PATH.'GestionDataBase.php'); 
+                include(MODEL_PATH.'Tarea.php'); 
 
-        header('location:../controllers/procesarListaTareas.php');
+                $id = $_GET['id'];
+                Tarea::deleteTarea($id);
+
+                header('location:../controllers/procesarListaTareas.php');
+
+                
+        }else if(($_SESSION['rol'] == 'Operario')){
+
+                header('location:procesarListaTareas.php');
+        
+        }else{
+        
+                header('location:procesarLogin.php');
+        
+        }
